Changed package name to com.elevatelab.ontimepro
