+++ OnTimePro/app/build.gradle
@@ -5,7 +5,7 @@ android {
     buildToolsVersion '30.0.2'
 
     defaultConfig {
-        applicationId "com.example.ontimepro"
+        applicationId "com.elevatelab.ontimepro"
         minSdkVersion 21
         targetSdkVersion 30
         versionCode 1
