+++ OnTimePro/app/src/androidTest/java/com/elevatelab/ontimepro/ExampleInstrumentedTest.java
@@ -1,14 +1,14 @@
-package com.example.ontimepro;
+package com.elevatelab.ontimepro;
 
 import android.content.Context;
 
-import androidx.test.platform.app.InstrumentationRegistry;
 import androidx.test.ext.junit.runners.AndroidJUnit4;
+import androidx.test.platform.app.InstrumentationRegistry;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
-import static org.junit.Assert.*;
+import static org.junit.Assert.assertEquals;
 
 /**
  * Instrumented test, which will execute on an Android device.
