+++ OnTimePro/app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.example.ontimepro">
+    package="com.elevatelab.ontimepro">
 
     <application
         android:allowBackup="true"
@@ -10,7 +10,7 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
         <activity
-            android:name=".SplashActivity"
+            android:name="com.elevatelab.ontimepro.SplashActivity"
             android:theme="@style/SplashTheme">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -18,10 +18,10 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".WelcomeActivity" />
-        <activity android:name=".SignUpActivity" />
-        <activity android:name=".LoginActivity" />
-        <activity android:name=".MainActivity" />
+        <activity android:name="com.elevatelab.ontimepro.WelcomeActivity" />
+        <activity android:name="com.elevatelab.ontimepro.SignUpActivity" />
+        <activity android:name="com.elevatelab.ontimepro.LoginActivity" />
+        <activity android:name="com.elevatelab.ontimepro.MainActivity" />
 
         <meta-data
             android:name="preloaded_fonts"
