+++ OnTimePro/app/src/main/java/com/elevatelab/ontimepro/MainActivity.java
@@ -1,0 +1,14 @@
+package com.elevatelab.ontimepro;
+
+import android.os.Bundle;
+
+import androidx.appcompat.app.AppCompatActivity;
+
+public class MainActivity extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_main);
+    }
+}
