+++ OnTimePro/app/src/main/java/com/elevatelab/ontimepro/SignUpActivity.java
@@ -1,14 +1,14 @@
-package com.example.ontimepro;
+package com.elevatelab.ontimepro;
 
 import android.os.Bundle;
 
 import androidx.appcompat.app.AppCompatActivity;
 
-public class LoginActivity extends AppCompatActivity {
+public class SignUpActivity extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_login);
+        setContentView(R.layout.activity_sign_up);
     }
 }
