+++ OnTimePro/app/src/main/java/com/elevatelab/ontimepro/SplashActivity.java
@@ -1,13 +1,10 @@
-package com.example.ontimepro;
+package com.elevatelab.ontimepro;
 
-import androidx.appcompat.app.AppCompatActivity;
-
-import android.content.ClipData;
 import android.content.Intent;
 import android.os.Bundle;
 import android.os.Handler;
-import android.view.WindowManager;
-import android.view.animation.AnimationUtils;
+
+import androidx.appcompat.app.AppCompatActivity;
 
 public class SplashActivity extends AppCompatActivity {
