+++ OnTimePro/app/src/main/java/com/elevatelab/ontimepro/WelcomeActivity.java
@@ -1,4 +1,4 @@
-package com.example.ontimepro;
+package com.elevatelab.ontimepro;
 
 import android.os.Bundle;
