+++ OnTimePro/app/src/test/java/com/elevatelab/ontimepro/ExampleUnitTest.java
@@ -1,8 +1,8 @@
-package com.example.ontimepro;
+package com.elevatelab.ontimepro;
 
 import org.junit.Test;
 
-import static org.junit.Assert.*;
+import static org.junit.Assert.assertEquals;
 
 /**
  * Example local unit test, which will execute on the development machine (host).
